requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-telegram-bot==20.3

